Space conversion expects only Real dimensions
1. It should be uniform in the perspective of skopt.
This is because we don't want skopt to mess with our stuff, since
we may already do. A normalization by skopt's side is harmless though.

2. Replace the hack with a better way to ensure that the
inclusive upper bound of skopt dimensions does not violate the
exclusive upper bound of Oríon's dimensions.

3. Register all hyperparameters to `BaseAlgorithm`. Make sure
that optimizer get the post initialization transformed space.

**XXX related to (3)**

This should be changed by the `PrimaryAlgo` and `BaseAlgorithm` classes;
space in the initialization of a concrete algorithm should be already
the transformed one, for algorithm developer's convenience and as this
is expected to be an often case when wrapping algorithm, the code
smells. Fix is possible since `BaseAlgorithm.requires` is a class
attribute => transformed space can actually be constructed before
concrete algorithm instantiation!