Make `optimizer` a lazy property
Why:

The transformed space is only available after the optimizer is fully
initialized. For this reason, we cannot build the optimizer inside
`init`.

Note:

`seed_rng` is called at the end of `init`. Since the optimizer cannot be
built at this time, we make the seeding lazy as well and delay its
application to right after the optimizer's creation.