Make the initialization of BO lazy
We need the transformed space to initialize the BO, but only the
original space is passed at construction time. The primary algo needs
the built algo to get the requirements that will be used to create the transformed space.
After that the transformed space is assigned to the algo. This is when
the lazy initialization will occur.