Adjust to new test suite
Why:

A new test suite for algorithms is available in orion.testing making it
easier to test all algorithms on the same grounds.

Note:

A few changes were needed to comply with the new tests suite.

- The random phase is implemented using Orion's space sampling instead
of relying on skopts samplers. This is to comply with transformed space,
particularly with cardinality. This is a drawback because we loose
access to various samplers in skopt, but there samplers where not
accessing with currently options of the wrapper.
- The method is_done is reimplemented to take account of original space
(not transformed one) cardinality.