Add documentation
Documentation should be included in the plugin instead of in the core
repo. The local documentation in docs/doc.rst will be included in the
global doc using the custom sphinx extension plugin-include, which
will fetch the doc.rst from latest release of the plugin.

But doing so we need to make a special test in the plugin repo
by fetching the whole documentation and including the local version to
make sure we are not breaking the global documentation.