Adapt to orion v0.2.2
The parallel strategy is now outside of Oríon's producer and must be
integrated inside the algorithms.

The initial points are now sampled from skopt's optimizer
to simplify the implementation and stay as close as possible to skopt.
Whenever a point sampled is equivalent to one already sampled, it is
added to a list of points sampled by the optimizer but not returned as a
suggested trial. The optimizer is recreated every time suggest is called
and saved points are used to set back its state. This is done to avoid
passing the same points to optimizer.tell multiple times if they are not
completed. The non completed points are assigned fake results using the
parallel strategy and passed to optimizer.tell(), but doing so multiple
times (which happens if the trials are not completed during multiple
calls to suggest()) will lead to having the GP model fitting the same
point multiple time with different results.